+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Adminer dev
 - Pretty print JSON in edit
+- Support multiline generated values in alter table
 - PostgreSQL: Offer foreign keys in create table
 - PostgreSQL: Add missing parentheses to CHECK export
 - PostgreSQL: Allow creating NOT DEFERRABLE foreign keys